Extract review CAD pipeline
